moving towards a  0.5 and a better tested app
