+++ GatewayToS3Function/GatewayToS3Function.csproj
@@ -8,17 +8,17 @@
     <NoWarn>CS1998</NoWarn>
   </PropertyGroup>
   <ItemGroup>
-    <PackageReference Include="Amazon.Lambda.Core" Version="1.0.0" />
-    <PackageReference Include="Amazon.Lambda.Serialization.Json" Version="1.4.0" />
+    <PackageReference Include="Amazon.Lambda.Core" Version="1.1.0" />
+    <PackageReference Include="Amazon.Lambda.Serialization.Json" Version="1.5.0" />
     <PackageReference Include="coverlet.msbuild" Version="2.5.1">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
     </PackageReference>
     <PackageReference Include="MindTouch.LambdaSharp" Version="0.4.0" />
     <PackageReference Include="Newtonsoft.Json" Version="12.0.1" />
-    <PackageReference Include="AWSSDK.Core" Version="3.3.31.3" />
-    <PackageReference Include="AWSSDK.Lambda" Version="3.3.19.12" />
-    <PackageReference Include="AWSSDK.S3" Version="3.3.31.11" />
+    <PackageReference Include="AWSSDK.Core" Version="3.3.31.7" />
+    <PackageReference Include="AWSSDK.Lambda" Version="3.3.19.16" />
+    <PackageReference Include="AWSSDK.S3" Version="3.3.31.15" />
   </ItemGroup>
   <ItemGroup>
     <DotNetCliToolReference Include="Amazon.Lambda.Tools" Version="2.2.0" />
