+++ Module.yml
@@ -1,60 +1,47 @@
-# The Module Name is used as prefix for all contained resources.
-Module: SecureFileUploadGateway
+Module: Owner.SecureFileUploadGateway
 
-# The Module Version is shown in the CloudFormation stack and Lambda function descriptions.
-Version: 1.3
+Version: 1.36
 
-# The Module Description is shown in the AWS CloudFormation console with the version number.
-Description: Resources allocated for secure file uploads to s3 from api gateway
-
-# The Variables section defines values and resources for the module.
-Variables:
-  - Var: SecureFileUploadGatewayBucket
+Items:
+  - Resource: SecureFileUploadGatewayBucket
     Scope: "*"
     Description: Default bucket of file uploads from api gateway
-    Resource:
-      Type: AWS::S3::Bucket
-      Allow: ReadWrite
+    Type: AWS::S3::Bucket
+    Allow: ReadWrite
 
-  - Var: CustomerUsagePlan
+  - Resource: CustomerUsagePlan
     Description: Limit usage for REST API
-    Resource:
-      Type: AWS::ApiGateway::UsagePlan
-      Properties:
-        UsagePlanName: CustomerUsagePlan
-        ApiStages:
-          - ApiId: !Ref Module::RestApi
-            Stage: !Ref Module::RestApiStage
-        Quota:
-          Limit: 100000
-          Period: MONTH
+    Type: AWS::ApiGateway::UsagePlan
+    Properties:
+      UsagePlanName: CustomerUsagePlan
+      ApiStages:
+        - ApiId: !Ref Module::RestApi
+          Stage: !Ref Module::RestApi::Stage
+      Quota:
+        Limit: 100000
+        Period: MONTH
 
     # https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-apigateway-apikey.html
-  - Var: Customer01ApiKey
+  - Resource: Customer01ApiKey
     Description: Limit access for customer01
-    Resource:
-      Type: AWS::ApiGateway::ApiKey
-      Properties:
-        Description: ApiKey for file uploader
-        Enabled: true
+    Type: AWS::ApiGateway::ApiKey
+    Properties:
+      Description: ApiKey for file uploader
+      Enabled: true
 
-#  - Var: AddApiKey
-#    Description: Add to usage plan
-#    Resource:
-#      Type: AWS::ApiGateway::UsagePlanKey
-#      Properties:
-#        KeyId: !Ref Module::RestApi
-#        KeyType: API_KEY
-#        UsagePlanId: !Ref CustomerUsagePlan # 0.5 fixes this
+  #  - Resource: AddApiKey
+  #    Description: Add to usage plan
+  #    Type: AWS::ApiGateway::UsagePlanKey
+  #    Properties:
+  #        KeyId: !Ref Module::RestApi
+  #        KeyType: API_KEY
+  #        UsagePlanId: !Ref CustomerUsagePlan # 0.5 fixes this
 
-Functions:
   - Function: GatewayToS3Function
-    Description: Puts a file to S3 from an Api Gateway request
+    Description: Post a file to S3 from an Api Gateway request
     Memory: 256
     Timeout: 60
     Sources:
       - Api: POST:/FileUpload
         ApiKeyRequired: true
         OperationName: Endpoint for file uploads
-        # Integration: #https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-apitgateway-method-integration.html
-        #  ContentHandling: CONVERT_TO_BINARY
