+++ TODO.md
@@ -1,13 +1,22 @@
 # To Do
 
+- [] move process logic to new class with different unit tests
 - [] key path for bucket
+- [] support multipart messages
+- [] dummy website to upload files
 - [] request for different bucket
 - [] cloud watch dashboard
-- [] api key usage and assignment. Ask about this. Black magic!!
-- [] support binary types in request. Base64 encoded.
-- [] Write only access to S3 bucket. Ask about this. https://github.com/LambdaSharp/LambdaSharpTool/blob/7e66c264ad4faa6a7695fd0e9facd28ed701dd33/src/MindTouch.LambdaSharp.Tool/Resources/IAM-Mappings.yml
+- [] api key usage and assignment.
+- [] Write only access to S3 bucket.
+- [x] support binary types in request. Base64 encoded.
 - [x] File name in request.
 - [x] Describe request in readme
 - [x] coverage report
 - [x] File upload for binary files.
 - [x] xunit testing for the basics
+
+## Other items
+
+- This still needs to be figured out
+  Integration: #https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-properties-apitgateway-method-integration.html
+  ContentHandling: CONVERT_TO_BINARY
